requirements.txt
awsebcli==3.7.4
backports.ssl-match-hostname==3.5.0.1
blessed==1.9.5
botocore==1.4.7
cement==2.4.0
colorama==0.3.3
Django==1.9.2
djangorestframework==3.3.2
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
eulfedora==1.5.2
eulxml==1.0.1
future==0.15.2
html5lib==0.9999999
isodate==0.5.4
jmespath==0.9.0
keepalive==0.5
lxml==3.6.0
MySQL-python==1.2.5
Naked==0.1.31
oauth==1.0.1
oauthlib==1.0.3
pathspec==0.3.3
Pillow==3.1.1
ply==3.8
progressbar==2.3
progressbar2==3.6.2
py==1.4.31
pycrypto==2.6.1
pyparsing==2.1.0
PyPDF2==1.25.1
pytest==2.9.1
python-dateutil==2.5.2
PyYAML==3.11
rdflib==4.2.1
requests==2.6.2
requests-toolbelt==0.6.0
semantic-version==2.5.0
shellescape==3.4.1
six==1.10.0
SPARQLWrapper==1.7.6
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.35.0
django-secure==1.0.1
